Fix build with --no-default-features and test this in github actions

+++ examples/unlink.rs
@@ -3,13 +3,15 @@
 // in the LICENSE file.
 
 fn main() {
-    let path = std::path::Path::new("/tmp/nc.unlink");
-    let ret = unsafe { nc::unlinkat(nc::AT_FDCWD, path, 0) };
-    assert!(ret.is_ok());
+    if cfg!(feature = "std") {
+        let path = std::path::Path::new("/tmp/nc.unlink");
+        let ret = unsafe { nc::unlinkat(nc::AT_FDCWD, path, 0) };
+        assert!(ret.is_ok());
 
-    let path = std::path::PathBuf::from("/tmp/nc.unlink");
-    let ret = unsafe { nc::unlinkat(nc::AT_FDCWD, path, 0) };
-    assert!(ret.is_ok());
+        let path = std::path::PathBuf::from("/tmp/nc.unlink");
+        let ret = unsafe { nc::unlinkat(nc::AT_FDCWD, path, 0) };
+        assert!(ret.is_ok());
+    }
 
     let path = "/tmp/nc.unlink";
     let ret = unsafe { nc::unlinkat(nc::AT_FDCWD, path, 0) };
